+++ test/libsolidity/syntaxTests/userDefinedValueType/explicit_conversions_wrap.sol
@@ -1,0 +1,8 @@
+type MyUint is uint;
+type MyAddress is address;
+
+function f() pure {
+    MyUint.wrap(5);
+    MyAddress.wrap(address(5));
+}
+// ----