+++ test/libsolidity/syntaxTests/userDefinedValueType/implicit_conversion.sol
@@ -1,0 +1,14 @@
+type MyInt is uint;
+type MyAddress is address;
+function f() pure {
+    MyInt a;
+    MyInt b = a;
+    MyAddress c;
+    MyAddress d = c;
+    b;
+    d;
+}
+
+function g(MyInt a) pure returns (MyInt) {
+    return a;
+}