+++ test/libsolidity/syntaxTests/userDefinedValueType/name_conflict_contract_warning.sol
@@ -1,0 +1,6 @@
+type MyAddress is address;
+contract C {
+    type MyAddress is address;
+}
+// ----
+// Warning 2519: (44-70): This declaration shadows an existing declaration.