+++ test/libsolidity/syntaxTests/userDefinedValueType/non_value_type_bytes.sol
@@ -1,0 +1,3 @@
+type MyBytes is bytes;
+// ----
+// TypeError 8129: (0-22): The underlying type of the user defined value type "MyBytes" is not a value type.