+++ test/libsolidity/syntaxTests/userDefinedValueType/non_value_type_mapping_err.sol
@@ -1,0 +1,3 @@
+type MyInt is mapping(uint => uint);
+// ----
+// TypeError 8657: (14-35): The underlying type for a user defined value type has to be an elementary value type.