+++ Changelog.md
@@ -3,6 +3,7 @@
 Language Features:
  * Inheritance: A function that overrides only a single interface function does not require the ``override`` specifier.
  * Type System: Support ``type().min`` and ``type().max`` for enums.
+ * User Defined Value Type: allows creating a zero cost abstraction over a value type with stricter type requirements.
 
 
 Compiler Features: